northd: add drop-debug-mode to add explicit drops
Add a new config flag called "drop-debug-mode" that makes northd add an
explicit default drop to all tables that currently do not have a default
(prio=0, match=1) lflow.

In the controller side, also add explicit default drop rules on physical
tables that need it.

When this mode is enabled the explicit drop actions  make it easier to
debug when OVN is dropping a packet.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>